# Admin-a-main/Admin-a-main/lib/api-spec/openapi.yaml
openapi: "3.0.3"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /pricing:
    get:
      operationId: getPricing
      tags: [pricing]
      summary: Get current pricing settings
      responses:
        "200":
          description: Pricing settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingSettings"
    patch:
      operationId: updatePricing
      tags: [pricing]
      summary: Update pricing settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingUpdate"
      responses:
        "200":
          description: Updated pricing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PricingSettings"
  /pricing/surge-zones:
    get:
      operationId: listSurgeZones
      tags: [pricing]
      summary: List surge pricing zones
      responses:
        "200":
          description: Surge zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SurgeZone"
    post:
      operationId: createSurgeZone
      tags: [pricing]
      summary: Create surge pricing zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurgeZoneInput"
      responses:
        "201":
          description: Created surge zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurgeZone"
  /pricing/surge-zones/{id}:
    patch:
      operationId: updateSurgeZone
      tags: [pricing]
      summary: Update surge zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SurgeZoneUpdate"
      responses:
        "200":
          description: Updated surge zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurgeZone"
    delete:
      operationId: deleteSurgeZone
      tags: [pricing]
      summary: Delete surge zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /service-areas:
    get:
      operationId: listServiceAreas
      tags: [settings]
      summary: List service areas
      responses:
        "200":
          description: Service areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceArea"
    post:
      operationId: createServiceArea
      tags: [settings]
      summary: Create service area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAreaInput"
      responses:
        "201":
          description: Created service area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceArea"
  /service-areas/{id}:
    patch:
      operationId: updateServiceArea
      tags: [settings]
      summary: Update service area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceAreaUpdate"
      responses:
        "200":
          description: Updated service area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceArea"
    delete:
      operationId: deleteServiceArea
      tags: [settings]
      summary: Delete service area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get app settings
      responses:
        "200":
          description: App settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
  /content/banners:
    get:
      operationId: listBanners
      tags: [content]
      summary: List banners
      responses:
        "200":
          description: Banners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Banner"
    post:
      operationId: createBanner
      tags: [content]
      summary: Create banner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerInput"
      responses:
        "201":
          description: Created banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
  /content/banners/{id}:
    patch:
      operationId: updateBanner
      tags: [content]
      summary: Update banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BannerUpdate"
      responses:
        "200":
          description: Updated banner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Banner"
    delete:
      operationId: deleteBanner
      tags: [content]
      summary: Delete banner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /content/announcements:
    get:
      operationId: listAnnouncements
      tags: [content]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [content]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /content/announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [content]
      summary: Update announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [content]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    PricingSettings:
      type: object
      required:
        - id
        - basePerKm
        - minimumFare
        - maximumFare
        - nightChargeMultiplier
        - nightChargeStartHour
        - nightChargeEndHour
        - waitingChargePerMin
        - cancellationFee
        - serviceRadiusKm
        - updatedAt
      properties:
        id:
          type: string
        basePerKm:
          type: number
        minimumFare:
          type: number
        maximumFare:
          type: number
        nightChargeMultiplier:
          type: number
        nightChargeStartHour:
          type: integer
        nightChargeEndHour:
          type: integer
        waitingChargePerMin:
          type: number
        cancellationFee:
          type: number
        serviceRadiusKm:
          type: number
        updatedAt:
          type: string
          format: date-time
    PricingUpdate:
      type: object
      properties:
        basePerKm:
          type: number
        minimumFare:
          type: number
        maximumFare:
          type: number
        nightChargeMultiplier:
          type: number
        nightChargeStartHour:
          type: integer
        nightChargeEndHour:
          type: integer
        waitingChargePerMin:
          type: number
        cancellationFee:
          type: number
        serviceRadiusKm:
          type: number
    SurgeZone:
      type: object
      required: [id, name, multiplier, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        multiplier:
          type: number
        startTime:
          type: string
        endTime:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SurgeZoneInput:
      type: object
      required: [name, multiplier]
      properties:
        name:
          type: string
        multiplier:
          type: number
        startTime:
          type: string
        endTime:
          type: string
        isActive:
          type: boolean
    SurgeZoneUpdate:
      type: object
      properties:
        name:
          type: string
        multiplier:
          type: number
        startTime:
          type: string
        endTime:
          type: string
        isActive:
          type: boolean
    ServiceArea:
      type: object
      required: [id, name, radiusKm, centerLat, centerLng, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        radiusKm:
          type: number
        centerLat:
          type: number
        centerLng:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ServiceAreaInput:
      type: object
      required: [name, radiusKm, centerLat, centerLng]
      properties:
        name:
          type: string
        radiusKm:
          type: number
        centerLat:
          type: number
        centerLng:
          type: number
        isActive:
          type: boolean
    ServiceAreaUpdate:
      type: object
      properties:
        name:
          type: string
        radiusKm:
          type: number
        centerLat:
          type: number
        centerLng:
          type: number
        isActive:
          type: boolean
...
    AppSettings:
      type: object
      required: [id, appName, supportNumber, commissionPercent, updatedAt]
      properties:
        id:
          type: string
        appName:
          type: string
        appLogoUrl:
          type: string
        supportNumber:
          type: string
        commissionPercent:
          type: number
        maintenanceMode:
          type: boolean
        registrationEnabled:
          type: boolean
        privacyPolicyUrl:
          type: string
        termsUrl:
          type: string
        updatedAt:
          type: string
          format: date-time
    AppSettingsUpdate:
      type: object
      properties:
        appName:
          type: string
        appLogoUrl:
          type: string
        supportNumber:
          type: string
        commissionPercent:
          type: number
        maintenanceMode:
          type: boolean
        registrationEnabled:
          type: boolean
        privacyPolicyUrl:
          type: string
        termsUrl:
          type: string
...
    Banner:
      type: object
      required: [id, title, imageUrl, isActive, order, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        order:
          type: integer
        createdAt:
          type: string
          format: date-time
    BannerInput:
      type: object
      required: [title, imageUrl]
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        order:
          type: integer
    BannerUpdate:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
        linkUrl:
          type: string
        isActive:
          type: boolean
        order:
          type: integer
    Announcement:
      type: object
      required: [id, title, content, isActive, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time